+++ src/blocker/integration.rs
@@ -1,14 +1,25 @@
+//! Integration with issue files and the `open` module.
+//!
+//! This module provides bridges for working with blockers embedded in issue files:
+//! - `set`: Set the current blocker source (issue file or standalone file)
+//! - `open`: Open the blocker source for editing
+//! - Extract/update blockers sections within issue files
+//!
+//! The `--integrated` flag on `set` and `open` commands enables use of issue files
+//! as the blocker source, using the `open/` module's file management and sync mechanics.
+
 use std::path::{Path, PathBuf};
 
 use clap::{Args, Subcommand};
 use color_eyre::eyre::{Result, bail, eyre};
 use todo::Extension;
 
-use crate::{
-	blocker::{self, LineType},
-	config::LiveSettings,
-	marker::Marker,
+use super::{
+	LineType,
+	operations::BlockerSequence,
+	standard::{classify_line, parse_parent_headers, strip_blocker_prefix},
 };
+use crate::{config::LiveSettings, marker::Marker};
 
 /// Returns the base directory for issue storage: XDG_DATA_HOME/todo/issues/
 fn issues_dir() -> PathBuf {
@@ -19,13 +30,13 @@ fn issues_dir() -> PathBuf {
 static CURRENT_BLOCKER_ISSUE_CACHE: &str = "current_blocker_issue.txt";
 
 #[derive(Args, Clone, Debug)]
-pub struct BlockerRewriteArgs {
+pub struct IntegrationArgs {
 	#[command(subcommand)]
-	command: Command,
+	pub command: IntegrationCommand,
 }
 
 #[derive(Clone, Debug, Subcommand)]
-pub enum Command {
+pub enum IntegrationCommand {
 	/// Set the current blocker file (uses same matching as `open`)
 	Set {
 		/// Pattern to match issue file (number, title, owner/repo pattern)
@@ -184,241 +195,6 @@ fn choose_issue_with_fzf(matches: &[PathBuf], initial_query: &str) -> Result<Opt
 	}
 }
 
-/// Extract blockers section from an issue file.
-/// Looks for `<!--blockers-->` marker (md) or `// blockers` (typst) in the body.
-/// Returns the content from that marker up to either:
-/// - End of body (before sub-issues or comments)
-/// - End of file
-///
-/// The returned content uses the same format as blocker.rs files.
-fn extract_blockers_section(content: &str) -> Option<String> {
-	let normalized = normalize_issue_indentation(content);
-	let lines: Vec<&str> = normalized.lines().collect();
-
-	// Find the blockers marker line
-	let mut blockers_start_idx = None;
-	let mut body_end_idx = lines.len();
-
-	for (idx, line) in lines.iter().enumerate() {
-		// Skip the issue title (first line)
-		if idx == 0 {
-			continue;
-		}
-
-		let stripped = line.strip_prefix('\t').unwrap_or(line);
-
-		// Check for blockers marker (must be in body, so indented)
-		if line.starts_with('\t') && blockers_start_idx.is_none() && matches!(Marker::decode(stripped, Extension::Md), Some(Marker::BlockersSection(_))) {
-			blockers_start_idx = Some(idx + 1); // Start from line after marker
-			continue;
-		}
-
-		// Check for sub-issue - any `- [ ]` or `- [x]` at same indent level (one tab) ends blockers
-		// This catches both marked sub-issues and newly added ones without markers
-		if line.starts_with('\t') && !line.starts_with("\t\t") && blockers_start_idx.is_some() {
-			let trimmed = stripped.trim();
-			if trimmed.starts_with("- [ ] ") || trimmed.starts_with("- [x] ") || trimmed.starts_with("- [X] ") {
-				body_end_idx = idx;
-				break;
-			}
-		}
-
-		// Check for comment marker - this also ends the body
-		if matches!(Marker::decode(stripped, Extension::Md), Some(Marker::Comment { .. } | Marker::NewComment)) {
-			body_end_idx = idx;
-			break;
-		}
-	}
-
-	// If no blockers marker found, return None
-	let start_idx = blockers_start_idx?;
-
-	// Extract lines from blockers marker to body end
-	// Remove the leading tab (body indent) from each line
-	let blockers_lines: Vec<&str> = lines[start_idx..body_end_idx].iter().map(|l| l.strip_prefix('\t').unwrap_or(l)).collect();
-
-	if blockers_lines.is_empty() {
-		return None;
-	}
-
-	Some(blockers_lines.join("\n"))
-}
-
-/// Get the current blocker from the blockers section.
-/// Uses the same logic as blocker.rs: last non-comment content line.
-fn get_current_blocker_from_content(blockers_content: &str) -> Option<String> {
-	blockers_content
-		.lines()
-		// Skip comment lines (tab-indented) - only consider content lines
-		.rfind(|s: &&str| !s.is_empty() && !s.starts_with('\t'))
-		.map(|s| s.to_owned())
-}
-
-/// Get the current blocker with parent headers prepended.
-/// Uses blocker.rs logic for parsing headers.
-/// If `fully_qualified` is true and `issue_path` is provided, prepends the issue title.
-fn get_current_blocker_with_headers(blockers_content: &str, fully_qualified: bool, issue_path: Option<&Path>) -> Option<String> {
-	let current = get_current_blocker_from_content(blockers_content)?;
-	let stripped = blocker::strip_blocker_prefix(&current);
-
-	let parent_headers = blocker::parse_parent_headers(blockers_content, &current);
-
-	// Build final output with parent headers
-	let mut parts = Vec::new();
-
-	// Add project name if fully_qualified is true
-	if fully_qualified && let Some(path) = issue_path {
-		// Extract project name from issue file path (filename without extension)
-		let project_name = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
-		if !project_name.is_empty() {
-			parts.push(project_name.to_string());
-		}
-	}
-
-	// Add parent headers
-	parts.extend(parent_headers);
-
-	// Add the stripped task
-	if parts.is_empty() {
-		Some(stripped.to_string())
-	} else {
-		Some(format!("{}: {stripped}", parts.join(": ")))
-	}
-}
-
-/// Pop the last content line from the blockers section.
-/// Returns the modified blockers content.
-fn pop_last_blocker(blockers_content: &str) -> String {
-	let lines: Vec<&str> = blockers_content.lines().collect();
-	let mut content_line_indices: Vec<usize> = Vec::new();
-
-	// Find indices of all content lines (headers and items, not comments)
-	for (idx, line) in lines.iter().enumerate() {
-		if let Some(line_type) = blocker::classify_line(line)
-			&& line_type.is_content()
-		{
-			content_line_indices.push(idx);
-		}
-	}
-
-	// Remove the last content line
-	if let Some(&last_content_idx) = content_line_indices.last() {
-		// Keep lines before the last content line
-		let new_lines: Vec<&str> = lines.iter().enumerate().filter(|(idx, _)| *idx < last_content_idx).map(|(_, line)| *line).collect();
-
-		new_lines.join("\n")
-	} else {
-		// No content lines to remove
-		blockers_content.to_string()
-	}
-}
-
-/// Update the blockers section in an issue file.
-/// Replaces the content between <!--blockers--> marker and the next section marker.
-/// Returns the updated full file content.
-fn update_blockers_in_issue(full_content: &str, new_blockers_content: &str) -> Option<String> {
-	let normalized = normalize_issue_indentation(full_content);
-	let lines: Vec<&str> = normalized.lines().collect();
-
-	// Find the blockers marker line
-	let mut blockers_start_idx = None;
-	let mut blockers_end_idx = lines.len();
-
-	for (idx, line) in lines.iter().enumerate() {
-		// Skip the issue title (first line)
-		if idx == 0 {
-			continue;
-		}
-
-		let stripped = line.strip_prefix('\t').unwrap_or(line);
-
-		// Check for blockers marker (must be in body, so indented)
-		if line.starts_with('\t') && blockers_start_idx.is_none() && matches!(Marker::decode(stripped, Extension::Md), Some(Marker::BlockersSection(_))) {
-			blockers_start_idx = Some(idx + 1); // Start from line after marker
-			continue;
-		}
-
-		// Check for sub-issue - any `- [ ]` or `- [x]` at same indent level (one tab) ends blockers
-		if line.starts_with('\t') && !line.starts_with("\t\t") && blockers_start_idx.is_some() {
-			let trimmed = stripped.trim();
-			if trimmed.starts_with("- [ ] ") || trimmed.starts_with("- [x] ") || trimmed.starts_with("- [X] ") {
-				blockers_end_idx = idx;
-				break;
-			}
-		}
-
-		// Check for comment marker - this also ends the body
-		if matches!(Marker::decode(stripped, Extension::Md), Some(Marker::Comment { .. } | Marker::NewComment)) {
-			blockers_end_idx = idx;
-			break;
-		}
-	}
-
-	// If no blockers marker found, return None
-	let start_idx = blockers_start_idx?;
-
-	// Build the new content
-	let mut result_lines: Vec<String> = Vec::new();
-
-	// Lines before blockers section
-	for line in &lines[..start_idx] {
-		result_lines.push(line.to_string());
-	}
-
-	// New blockers content (add body indent)
-	for line in new_blockers_content.lines() {
-		if line.is_empty() {
-			result_lines.push(String::new());
-		} else {
-			result_lines.push(format!("\t{line}"));
-		}
-	}
-
-	// Lines after blockers section
-	for line in &lines[blockers_end_idx..] {
-		result_lines.push(line.to_string());
-	}
-
-	Some(result_lines.join("\n"))
-}
-
-/// List all blockers from the content.
-/// Returns tuples of (text, is_header, is_completed).
-fn list_blockers_from_content(blockers_content: &str) -> Vec<(String, bool, bool)> {
-	let mut result = Vec::new();
-
-	for line in blockers_content.lines() {
-		// Skip empty lines and comments (tab-indented)
-		if line.is_empty() || line.starts_with('\t') {
-			continue;
-		}
-
-		let line_type = blocker::classify_line(line);
-		match line_type {
-			Some(LineType::Header { text, .. }) => {
-				result.push((text, true, false));
-			}
-			Some(LineType::Item) => {
-				// Check if it's a checkbox item
-				let trimmed = line.trim();
-				let (completed, text) = if let Some(rest) = trimmed.strip_prefix("- [x] ").or_else(|| trimmed.strip_prefix("- [X] ")) {
-					(true, rest.to_string())
-				} else if let Some(rest) = trimmed.strip_prefix("- [ ] ") {
-					(false, rest.to_string())
-				} else {
-					// Regular item (- prefix)
-					let text = blocker::strip_blocker_prefix(trimmed);
-					(false, text.to_string())
-				};
-				result.push((text, false, completed));
-			}
-			_ => {}
-		}
-	}
-
-	result
-}
-
 /// Normalize issue content by converting space-based indentation to tab-based.
 /// Same logic as open.rs
 fn normalize_issue_indentation(content: &str) -> String {
@@ -473,9 +249,206 @@ fn normalize_issue_indentation(content: &str) -> String {
 		.join("\n")
 }
 
-pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<()> {
+/// Extract blockers section from an issue file.
+/// Looks for `<!--blockers-->` marker (md) or `// blockers` (typst) in the body.
+/// Returns the content from that marker up to either:
+/// - End of body (before sub-issues or comments)
+/// - End of file
+///
+/// The returned content uses the same format as blocker.rs files.
+fn extract_blockers_section(content: &str) -> Option<BlockerSequence> {
+	let normalized = normalize_issue_indentation(content);
+	let lines: Vec<&str> = normalized.lines().collect();
+
+	// Find the blockers marker line
+	let mut blockers_start_idx = None;
+	let mut body_end_idx = lines.len();
+
+	for (idx, line) in lines.iter().enumerate() {
+		// Skip the issue title (first line)
+		if idx == 0 {
+			continue;
+		}
+
+		let stripped = line.strip_prefix('\t').unwrap_or(line);
+
+		// Check for blockers marker (must be in body, so indented)
+		if line.starts_with('\t') && blockers_start_idx.is_none() && matches!(Marker::decode(stripped, Extension::Md), Some(Marker::BlockersSection(_))) {
+			blockers_start_idx = Some(idx + 1); // Start from line after marker
+			continue;
+		}
+
+		// Check for sub-issue - any `- [ ]` or `- [x]` at same indent level (one tab) ends blockers
+		// This catches both marked sub-issues and newly added ones without markers
+		if line.starts_with('\t') && !line.starts_with("\t\t") && blockers_start_idx.is_some() {
+			let trimmed = stripped.trim();
+			if trimmed.starts_with("- [ ] ") || trimmed.starts_with("- [x] ") || trimmed.starts_with("- [X] ") {
+				body_end_idx = idx;
+				break;
+			}
+		}
+
+		// Check for comment marker - this also ends the body
+		if matches!(Marker::decode(stripped, Extension::Md), Some(Marker::Comment { .. } | Marker::NewComment)) {
+			body_end_idx = idx;
+			break;
+		}
+	}
+
+	// If no blockers marker found, return None
+	let start_idx = blockers_start_idx?;
+
+	// Extract lines from blockers marker to body end
+	// Remove the leading tab (body indent) from each line
+	let blockers_lines: Vec<&str> = lines[start_idx..body_end_idx].iter().map(|l| l.strip_prefix('\t').unwrap_or(l)).collect();
+
+	if blockers_lines.is_empty() {
+		return None;
+	}
+
+	Some(BlockerSequence::new(blockers_lines.join("\n")))
+}
+
+/// Get the current blocker with parent headers prepended.
+/// Uses blocker.rs logic for parsing headers.
+/// If `fully_qualified` is true and `issue_path` is provided, prepends the issue title.
+fn get_current_blocker_with_headers(blockers: &BlockerSequence, fully_qualified: bool, issue_path: Option<&Path>) -> Option<String> {
+	let current = blockers.current()?;
+	let stripped = strip_blocker_prefix(&current);
+
+	let parent_headers = parse_parent_headers(blockers.content(), &current);
+
+	// Build final output with parent headers
+	let mut parts = Vec::new();
+
+	// Add project name if fully_qualified is true
+	if fully_qualified && let Some(path) = issue_path {
+		// Extract project name from issue file path (filename without extension)
+		let project_name = path.file_stem().and_then(|s| s.to_str()).unwrap_or("");
+		if !project_name.is_empty() {
+			parts.push(project_name.to_string());
+		}
+	}
+
+	// Add parent headers
+	parts.extend(parent_headers);
+
+	// Add the stripped task
+	if parts.is_empty() {
+		Some(stripped.to_string())
+	} else {
+		Some(format!("{}: {stripped}", parts.join(": ")))
+	}
+}
+
+/// List all content lines (headers and items), returning tuples of (text, is_header, is_completed)
+fn list_blockers_from_content(blockers: &BlockerSequence) -> Vec<(String, bool, bool)> {
+	let mut result = Vec::new();
+
+	for line in blockers.content().lines() {
+		// Skip empty lines and comments (tab-indented)
+		if line.is_empty() || line.starts_with('\t') {
+			continue;
+		}
+
+		let line_type = classify_line(line);
+		match line_type {
+			Some(LineType::Header { text, .. }) => {
+				result.push((text, true, false));
+			}
+			Some(LineType::Item) => {
+				// Check if it's a checkbox item
+				let trimmed = line.trim();
+				let (completed, text) = if let Some(rest) = trimmed.strip_prefix("- [x] ").or_else(|| trimmed.strip_prefix("- [X] ")) {
+					(true, rest.to_string())
+				} else if let Some(rest) = trimmed.strip_prefix("- [ ] ") {
+					(false, rest.to_string())
+				} else {
+					// Regular item (- prefix)
+					let text = strip_blocker_prefix(trimmed);
+					(false, text.to_string())
+				};
+				result.push((text, false, completed));
+			}
+			_ => {}
+		}
+	}
+
+	result
+}
+
+/// Update the blockers section in an issue file.
+/// Replaces the content between <!--blockers--> marker and the next section marker.
+/// Returns the updated full file content.
+fn update_blockers_in_issue(full_content: &str, new_blockers: &BlockerSequence) -> Option<String> {
+	let normalized = normalize_issue_indentation(full_content);
+	let lines: Vec<&str> = normalized.lines().collect();
+
+	// Find the blockers marker line
+	let mut blockers_start_idx = None;
+	let mut blockers_end_idx = lines.len();
+
+	for (idx, line) in lines.iter().enumerate() {
+		// Skip the issue title (first line)
+		if idx == 0 {
+			continue;
+		}
+
+		let stripped = line.strip_prefix('\t').unwrap_or(line);
+
+		// Check for blockers marker (must be in body, so indented)
+		if line.starts_with('\t') && blockers_start_idx.is_none() && matches!(Marker::decode(stripped, Extension::Md), Some(Marker::BlockersSection(_))) {
+			blockers_start_idx = Some(idx + 1); // Start from line after marker
+			continue;
+		}
+
+		// Check for sub-issue - any `- [ ]` or `- [x]` at same indent level (one tab) ends blockers
+		if line.starts_with('\t') && !line.starts_with("\t\t") && blockers_start_idx.is_some() {
+			let trimmed = stripped.trim();
+			if trimmed.starts_with("- [ ] ") || trimmed.starts_with("- [x] ") || trimmed.starts_with("- [X] ") {
+				blockers_end_idx = idx;
+				break;
+			}
+		}
+
+		// Check for comment marker - this also ends the body
+		if matches!(Marker::decode(stripped, Extension::Md), Some(Marker::Comment { .. } | Marker::NewComment)) {
+			blockers_end_idx = idx;
+			break;
+		}
+	}
+
+	// If no blockers marker found, return None
+	let start_idx = blockers_start_idx?;
+
+	// Build the new content
+	let mut result_lines: Vec<String> = Vec::new();
+
+	// Lines before blockers section
+	for line in &lines[..start_idx] {
+		result_lines.push(line.to_string());
+	}
+
+	// New blockers content (add body indent)
+	for line in new_blockers.content().lines() {
+		if line.is_empty() {
+			result_lines.push(String::new());
+		} else {
+			result_lines.push(format!("\t{line}"));
+		}
+	}
+
+	// Lines after blockers section
+	for line in &lines[blockers_end_idx..] {
+		result_lines.push(line.to_string());
+	}
+
+	Some(result_lines.join("\n"))
+}
+
+pub async fn main(_settings: &LiveSettings, args: IntegrationArgs) -> Result<()> {
 	match args.command {
-		Command::Set { pattern } => {
+		IntegrationCommand::Set { pattern } => {
 			// Use same matching logic as open command
 			let matches = search_issue_files(&pattern)?;
 
@@ -515,8 +488,8 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 
 			// Show current blocker if any
 			let content = std::fs::read_to_string(&issue_path)?;
-			if let Some(blockers_section) = extract_blockers_section(&content) {
-				if let Some(current) = get_current_blocker_with_headers(&blockers_section, false, Some(&issue_path)) {
+			if let Some(blockers) = extract_blockers_section(&content) {
+				if let Some(current) = get_current_blocker_with_headers(&blockers, false, Some(&issue_path)) {
 					println!("Current blocker: {current}");
 				} else {
 					println!("Blockers section is empty.");
@@ -526,18 +499,18 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 			}
 		}
 
-		Command::List => {
-			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-rewrite set <pattern>` first."))?;
+		IntegrationCommand::List => {
+			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-integrated set <pattern>` first."))?;
 
 			let content = std::fs::read_to_string(&issue_path)?;
 
-			if let Some(blockers_section) = extract_blockers_section(&content) {
-				let blockers = list_blockers_from_content(&blockers_section);
+			if let Some(blockers) = extract_blockers_section(&content) {
+				let items = list_blockers_from_content(&blockers);
 
-				if blockers.is_empty() {
+				if items.is_empty() {
 					println!("Blockers section is empty.");
 				} else {
-					for (text, is_header, completed) in &blockers {
+					for (text, is_header, completed) in &items {
 						if *is_header {
 							println!("# {text}");
 						} else {
@@ -551,13 +524,13 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 			}
 		}
 
-		Command::Current { fully_qualified } => {
-			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-rewrite set <pattern>` first."))?;
+		IntegrationCommand::Current { fully_qualified } => {
+			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-integrated set <pattern>` first."))?;
 
 			let content = std::fs::read_to_string(&issue_path)?;
 
-			if let Some(blockers_section) = extract_blockers_section(&content)
-				&& let Some(current) = get_current_blocker_with_headers(&blockers_section, fully_qualified, Some(&issue_path))
+			if let Some(blockers) = extract_blockers_section(&content)
+				&& let Some(current) = get_current_blocker_with_headers(&blockers, fully_qualified, Some(&issue_path))
 			{
 				const MAX_LEN: usize = 70;
 				match current.len() {
@@ -569,7 +542,7 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 			// No blockers section - silently exit
 		}
 
-		Command::Open { pattern, set_after } => {
+		IntegrationCommand::Open { pattern, set_after } => {
 			// Determine which file to open
 			let issue_path = if let Some(pat) = pattern {
 				// Use pattern matching to find issue file
@@ -598,7 +571,7 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 				}
 			} else {
 				// Use current blocker issue
-				get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-rewrite set <pattern>` first."))?
+				get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-integrated set <pattern>` first."))?
 			};
 
 			// Open the file with $EDITOR
@@ -619,29 +592,29 @@ pub async fn main(_settings: &LiveSettings, args: BlockerRewriteArgs) -> Result<
 			}
 		}
 
-		Command::Pop => {
-			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-rewrite set <pattern>` first."))?;
+		IntegrationCommand::Pop => {
+			let issue_path = get_current_blocker_issue().ok_or_else(|| eyre!("No blocker file set. Use `todo blocker-integrated set <pattern>` first."))?;
 
 			let content = std::fs::read_to_string(&issue_path)?;
 
-			if let Some(blockers_section) = extract_blockers_section(&content) {
+			if let Some(mut blockers) = extract_blockers_section(&content) {
 				// Get the current blocker before popping (for display)
-				let current_before = get_current_blocker_from_content(&blockers_section);
+				let current_before = blockers.current();
 
 				// Pop the last blocker
-				let new_blockers = pop_last_blocker(&blockers_section);
+				blockers.pop()?;
 
 				// Update the issue file with new blockers section
-				if let Some(updated_content) = update_blockers_in_issue(&content, &new_blockers) {
+				if let Some(updated_content) = update_blockers_in_issue(&content, &blockers) {
 					std::fs::write(&issue_path, updated_content)?;
 
 					if let Some(popped) = current_before {
-						let stripped = blocker::strip_blocker_prefix(&popped);
+						let stripped = strip_blocker_prefix(&popped);
 						println!("Popped: {stripped}");
 					}
 
 					// Show new current blocker
-					if let Some(new_current) = get_current_blocker_with_headers(&new_blockers, false, Some(&issue_path)) {
+					if let Some(new_current) = get_current_blocker_with_headers(&blockers, false, Some(&issue_path)) {
 						println!("Current: {new_current}");
 					} else {
 						println!("Blockers section is now empty.");
@@ -679,15 +652,15 @@ mod tests {
 	# Phase 2
 	- Third task
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		# Phase 1
-		- First task
-			comment on first task
-		- Second task
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r#"
+  # Phase 1
+  - First task
+  	comment on first task
+  - Second task
 
-		# Phase 2
-		- Third task
-		");
+  # Phase 2
+  - Third task
+  "#);
 	}
 
 	#[test]
@@ -700,10 +673,10 @@ mod tests {
 	# Main
 	- Do something
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		# Main
-		- Do something
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r#"
+  # Main
+  - Do something
+  "#);
 	}
 
 	#[test]
@@ -716,10 +689,10 @@ mod tests {
 	- Blocker two
 	- [ ] Sub-issue <!--sub https://github.com/owner/repo/issues/2-->
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		- Blocker one
-		- Blocker two
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r"
+  - Blocker one
+  - Blocker two
+  ");
 	}
 
 	#[test]
@@ -734,11 +707,11 @@ mod tests {
 	- [ ] new sub-issue I just added
 		without any markers around it
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		- last
-		- middle
-		- first
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r"
+  - last
+  - middle
+  - first
+  ");
 	}
 
 	#[test]
@@ -753,10 +726,10 @@ mod tests {
 	<!--https://github.com/owner/repo/issues/1#issuecomment-123-->
 	Comment body here.
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		- Blocker one
-		- Blocker two
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r"
+  - Blocker one
+  - Blocker two
+  ");
 	}
 
 	#[test]
@@ -770,38 +743,38 @@ mod tests {
 
 	#[test]
 	fn test_get_current_blocker() {
-		let blockers_content = "# Phase 1\n- First task\n\tcomment\n- Second task\n# Phase 2\n- Third task";
-		assert_snapshot!(get_current_blocker_from_content(blockers_content).unwrap(), @"- Third task");
+		let blockers = BlockerSequence::new("# Phase 1\n- First task\n\tcomment\n- Second task\n# Phase 2\n- Third task".to_string());
+		assert_snapshot!(blockers.current().unwrap(), @"- Third task");
 	}
 
 	#[test]
 	fn test_get_current_blocker_with_headers() {
-		let blockers_content = "# Phase 1\n- First task\n# Phase 2\n- Third task";
-		assert_snapshot!(get_current_blocker_with_headers(blockers_content, false, None).unwrap(), @"Phase 2: Third task");
+		let blockers = BlockerSequence::new("# Phase 1\n- First task\n# Phase 2\n- Third task".to_string());
+		assert_snapshot!(get_current_blocker_with_headers(&blockers, false, None).unwrap(), @"Phase 2: Third task");
 	}
 
 	#[test]
 	fn test_get_current_blocker_with_headers_fully_qualified() {
-		let blockers_content = "# Phase 1\n- First task\n# Phase 2\n- Third task";
+		let blockers = BlockerSequence::new("# Phase 1\n- First task\n# Phase 2\n- Third task".to_string());
 		let issue_path = Path::new("/home/user/issues/my_project.md");
-		assert_snapshot!(get_current_blocker_with_headers(blockers_content, true, Some(issue_path)).unwrap(), @"my_project: Phase 2: Third task");
+		assert_snapshot!(get_current_blocker_with_headers(&blockers, true, Some(issue_path)).unwrap(), @"my_project: Phase 2: Third task");
 	}
 
 	#[test]
 	fn test_list_blockers() {
-		let blockers_content = "# Phase 1\n- [x] Completed task\n- [ ] Pending task\n- Regular item";
-		let items = list_blockers_from_content(blockers_content);
+		let blockers = BlockerSequence::new("# Phase 1\n- [x] Completed task\n- [ ] Pending task\n- Regular item".to_string());
+		let items = list_blockers_from_content(&blockers);
 		assert_snapshot!(format!("{:?}", items), @r#"[("Phase 1", true, false), ("Completed task", false, true), ("Pending task", false, false), ("Regular item", false, false)]"#);
 	}
 
 	#[test]
 	fn test_normalize_indentation() {
 		let content = "- [ ] Title <!--url-->\n    Body with 4 spaces\n    - List item";
-		assert_snapshot!(normalize_issue_indentation(content), @"
-		- [ ] Title <!--url-->
-			Body with 4 spaces
-			- List item
-		");
+		assert_snapshot!(normalize_issue_indentation(content), @r"
+  - [ ] Title <!--url-->
+  	Body with 4 spaces
+  	- List item
+  ");
 	}
 
 	#[test]
@@ -813,10 +786,10 @@ mod tests {
 	- Task one
 	- Task two
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		- Task one
-		- Task two
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r"
+  - Task one
+  - Task two
+  ");
 	}
 
 	#[test]
@@ -829,10 +802,10 @@ mod tests {
 	- Task one
 	- Task two
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"
-		- Task one
-		- Task two
-		");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @r"
+  - Task one
+  - Task two
+  ");
 	}
 
 	#[test]
@@ -844,50 +817,48 @@ mod tests {
 	// blocker
 	- Task one
 "#;
-		assert_snapshot!(extract_blockers_section(content).unwrap(), @"- Task one");
+		assert_snapshot!(extract_blockers_section(content).unwrap().content(), @"- Task one");
 	}
 
 	#[test]
 	fn test_pop_last_blocker_simple() {
-		let blockers_content = "- First task\n- Second task\n- Third task";
-		assert_snapshot!(pop_last_blocker(blockers_content), @"
-		- First task
-		- Second task
-		");
+		let mut blockers = BlockerSequence::new("- First task\n- Second task\n- Third task".to_string());
+		blockers.pop().unwrap();
+		assert_snapshot!(blockers.content(), @r"
+  - First task
+  - Second task
+  ");
 	}
 
 	#[test]
 	fn test_pop_last_blocker_with_headers() {
-		let blockers_content = "# Phase 1\n- First task\n# Phase 2\n- Third task";
+		let mut blockers = BlockerSequence::new("# Phase 1\n- First task\n# Phase 2\n- Third task".to_string());
 		// Should pop "- Third task", leaving the header
-		assert_snapshot!(pop_last_blocker(blockers_content), @"
-		# Phase 1
-		- First task
-		# Phase 2
-		");
+		blockers.pop().unwrap();
+		assert_snapshot!(blockers.content(), @r"
+  # Phase 1
+  - First task
+
+  # Phase 2
+  ");
 	}
 
 	#[test]
 	fn test_pop_last_blocker_with_comments() {
-		let blockers_content = "- First task\n\tcomment on first\n- Second task\n\tcomment on second";
+		let mut blockers = BlockerSequence::new("- First task\n\tcomment on first\n- Second task\n\tcomment on second".to_string());
 		// Should pop "- Second task", removing its comment too (comments belong to content above)
-		assert_snapshot!(pop_last_blocker(blockers_content), @"
-		- First task
-			comment on first
-		");
+		blockers.pop().unwrap();
+		assert_snapshot!(blockers.content(), @r"
+  - First task
+  	comment on first
+  ");
 	}
 
 	#[test]
 	fn test_pop_last_blocker_empty() {
-		let blockers_content = "";
-		assert_snapshot!(pop_last_blocker(blockers_content), @"");
-	}
-
-	#[test]
-	fn test_pop_last_blocker_only_comments() {
-		let blockers_content = "\tcomment only";
-		// No content to pop
-		assert_snapshot!(pop_last_blocker(blockers_content), @"	comment only");
+		let mut blockers = BlockerSequence::empty();
+		blockers.pop().unwrap();
+		assert_snapshot!(blockers.content(), @"");
 	}
 
 	#[test]
@@ -900,15 +871,15 @@ mod tests {
 	- Second task
 	- Third task
 "#;
-		let new_blockers = "- First task\n- Second task";
-		assert_snapshot!(update_blockers_in_issue(content, new_blockers).unwrap(), @"
-		- [ ] Issue Title <!--https://github.com/owner/repo/issues/1-->
-			Body text.
+		let new_blockers = BlockerSequence::new("- First task\n- Second task".to_string());
+		assert_snapshot!(update_blockers_in_issue(content, &new_blockers).unwrap(), @r"
+  - [ ] Issue Title <!--https://github.com/owner/repo/issues/1-->
+  	Body text.
 
-			<!--blockers-->
-			- First task
-			- Second task
-		");
+  	<!--blockers-->
+  	- First task
+  	- Second task
+  ");
 	}
 
 	#[test]
@@ -921,18 +892,18 @@ mod tests {
 	- Second task
 	- [ ] Sub-issue <!--sub https://github.com/owner/repo/issues/2-->
 "#;
-		let new_blockers = "- First task";
-		let result = update_blockers_in_issue(content, new_blockers).unwrap();
+		let new_blockers = BlockerSequence::new("- First task".to_string());
+		let result = update_blockers_in_issue(content, &new_blockers).unwrap();
 		// Should preserve the sub-issue
 		assert!(result.contains("- [ ] Sub-issue"));
-		assert_snapshot!(result, @"
-		- [ ] Issue Title <!--https://github.com/owner/repo/issues/1-->
-			Body.
+		assert_snapshot!(result, @r"
+  - [ ] Issue Title <!--https://github.com/owner/repo/issues/1-->
+  	Body.
 
-			<!--blockers-->
-			- First task
-			- [ ] Sub-issue <!--sub https://github.com/owner/repo/issues/2-->
-		");
+  	<!--blockers-->
+  	- First task
+  	- [ ] Sub-issue <!--sub https://github.com/owner/repo/issues/2-->
+  ");
 	}
 
 	#[test]
@@ -940,6 +911,7 @@ mod tests {
 		let content = r#"- [ ] Issue Title <!--https://github.com/owner/repo/issues/1-->
 	Body without blockers marker.
 "#;
-		assert!(update_blockers_in_issue(content, "- Task").is_none());
+		let blockers = BlockerSequence::new("- Task".to_string());
+		assert!(update_blockers_in_issue(content, &blockers).is_none());
 	}
 }
