+++ src/main.rs
@@ -1,6 +1,5 @@
 #![allow(clippy::len_zero)]
 mod blocker;
-mod blocker_rewrite;
 mod clockify;
 pub mod config;
 pub mod error;
@@ -49,10 +48,10 @@ enum Commands {
 	Milestones(milestones::MilestonesArgs),
 	/// Shell aliases and hooks. Usage: `todos init <shell> | source`
 	Init(shell_init::ShellInitArgs),
-	/// Blockers tree
+	/// Blockers tree (standalone files)
 	Blocker(blocker::BlockerArgs),
-	/// Blocker management linked to issue files
-	BlockerRewrite(blocker_rewrite::BlockerRewriteArgs),
+	/// Blocker management integrated with issue files
+	BlockerIntegrated(blocker::IntegrationArgs),
 	/// Clockify time tracking
 	Clockify(clockify::ClockifyArgs),
 	/// Performance evaluation with screenshots
@@ -125,7 +124,7 @@ async fn main() {
 			Ok(())
 		}
 		Commands::Blocker(args) => blocker::main(&settings, args).await,
-		Commands::BlockerRewrite(args) => blocker_rewrite::main(&settings, args).await,
+		Commands::BlockerIntegrated(args) => blocker::main_integrated(&settings, args).await,
 		Commands::Clockify(args) => clockify::main(&settings, args).await,
 		Commands::PerfEval(args) => perf_eval::main(&settings, args).await,
 		Commands::WatchMonitors(args) => watch_monitors::main(&settings, args),
